refactor: 공통스타일 분리
