+++ client/src/components/TemplateList/HomeTemplate/HomeTemplate.tsx
@@ -1,10 +1,12 @@
 import { TemplateProps } from '../TemplateProps';
-import * as styles from './HomeTemplate.styles';
+import * as styles from '../Common.styles';
+import styled from '@emotion/styled';
+import { COLORS } from '@/styles/colors';
 
 const HomeTemplate = ({ title, time, place, memo, friends }: TemplateProps) => {
   return (
     <styles.Container>
-      <styles.InfoBox>
+      <InfoBox>
         <styles.TitleBox>{title}</styles.TitleBox>
         <styles.DateBox>{time}</styles.DateBox>
         <styles.PlaceBox>{place}</styles.PlaceBox>
@@ -14,10 +16,27 @@ const HomeTemplate = ({ title, time, place, memo, friends }: TemplateProps) => {
             return <styles.PersonInfo key={el}>{el}</styles.PersonInfo>;
           })}
         </styles.PeopleBox>
-      </styles.InfoBox>
+      </InfoBox>
       <styles.PlusBox>+</styles.PlusBox>
     </styles.Container>
   );
 };
+
+const InfoBox = styled.div`
+  width: 250px;
+  background-color: ${COLORS.grey7};
+  padding: 15px 20px;
+  border-radius: 10px;
+  margin-right: 12px;
+`;
+
+export const PlusBox = styled.button`
+  width: 50px;
+  background-color: ${COLORS.primary};
+  border-radius: 10px;
+  color: white;
+  font-size: 30px;
+  border: none;
+`;
 
 export default HomeTemplate;
