+++ client/src/components/TemplateList/TemplatePage/Template.tsx
@@ -1,10 +1,12 @@
 import { TemplateProps } from '../TemplateProps';
-import * as styles from './Template.styles';
+import * as styles from '../Common.styles';
+import { COLORS } from '@/styles/colors';
+import styled from '@emotion/styled';
 
 const Template = ({ title, time, place, memo, friends }: TemplateProps) => {
   return (
     <styles.Container>
-      <styles.InfoBox>
+      <InfoBox>
         <styles.TitleBox>{title}</styles.TitleBox>
         <styles.DateBox>{time}</styles.DateBox>
         <styles.PlaceBox>{place}</styles.PlaceBox>
@@ -14,10 +16,23 @@ const Template = ({ title, time, place, memo, friends }: TemplateProps) => {
             return <styles.PersonInfo key={el}>{el}</styles.PersonInfo>;
           })}
         </styles.PeopleBox>
-      </styles.InfoBox>
-      <styles.SideLine />
+      </InfoBox>
+      <SideLine />
     </styles.Container>
   );
 };
+
+const InfoBox = styled.div`
+  width: 329px;
+  background-color: ${COLORS.grey7};
+  padding: 15px 20px;
+  border-radius: 10px 0 0 10px;
+`;
+
+const SideLine = styled.div`
+  background-color: ${COLORS.primary};
+  border-radius: 0 10px 10px 0;
+  width: 5px;
+`;
 
 export default Template;
